requirements.txt
scipy
tortoise-tts
flask
flask-cors
torchaudio
transformers==4.31.0
einops==0.5.0
rotary_embedding_torch==0.1.5
unidecode==1.3.5
